Add 3D geometry elements Point3D and Vector3D
Point and vectors can be added and subtracted where it makes sense. See
test cases that cover these operations. Changed build to make a library
for the application code so that tests are easy to add to build.

I also found a very complete .gitignore file for C/C++ development.

Finally, I added a .clang-tidy configuration file that will
suppress clang-tidy from processing test files that would
otherwise generate a great deal of output because of the doctest
macros.